requirements.txt
mysql-connector
translate
colorama
pillow
schedule
smtplib
